+++ tests/test_load_queries.py
